requirements.txt
accumulation-tree==0.6.2
amqp==5.0.6
appdirs==1.4.4
asgiref==3.4.1
asn1ate==0.6.0
attrs==21.2.0
billiard==3.6.4.0
cached-property==1.5.2
celery==5.0.5
certifi==2021.5.30
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
Django==3.1.13
django-admin-rangefilter==0.5.4
django-health-check==3.14.3
django-model-utils==4.0.0
django-mysql==3.9.0
django-redis==4.12.1
django-seed==0.2.2
django-wkhtmltopdf==3.3.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
et-xmlfile==1.1.0
Faker==8.10.0
idna==2.10
isodate==0.6.0
jdatetime==3.6.2
jdcal==1.4.1
jsonfield==3.0.0
Khayyam==3.0.17
kombu==5.1.0
lxml==4.6.3
mysqlclient==2.0.3
openpyxl==2.6.4
pika==1.1.0
prompt-toolkit==3.0.19
pyasn1==0.4.8
pybrake==0.4.6
pydub==0.24.1
PyJWT==1.7.1
PyMySQL==0.10.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-magic==0.4.24
pytz==2020.5
pyudorandom==1.0.0
redis==3.5.3
requests==2.24.0
requests-file==1.5.1
requests-toolbelt==0.9.1
setproctitle==1.1.10
six==1.15.0
sqlparse==0.4.1
tdigest==0.5.2.2
text-unidecode==1.3
urllib3==1.25.11
uWSGI==2.0.19.1
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.57.0
wkhtmltopdf==0.2
xlrd==1.2.0
zeep==4.0.0
